Moved optoin in index.js
