+++ README.md
@@ -1,12 +1,23 @@
-![Command Line Parser](http://i.imgur.com/Gcn3nvW.png)
+<!---------------------------------------------------------------------------->
+<!-- STOP, LOOK & LISTEN!                                                   -->
+<!-- ====================                                                   -->
+<!-- Do NOT edit this file directly since it's generated from a template    -->
+<!-- file, using https://github.com/IonicaBizau/node-blah                   -->
+<!--                                                                        -->
+<!-- If you found a typo in documentation, fix it in the source files       -->
+<!-- (`lib/*.js`) and make a pull request.                                  -->
+<!--                                                                        -->
+<!-- If you have any other ideas, open an issue.                            -->
+<!--                                                                        -->
+<!-- Please consider reading the contribution steps (CONTRIBUTING.md).      -->
+<!-- * * * Thanks! * * *                                                    -->
+<!---------------------------------------------------------------------------->
 
-# clp
+![clp](http://i.imgur.com/Gcn3nvW.png)
+
+# clp [![Donate now][donate-now]][paypal-donations]
+
 A tiny and fast command line arguments parser and help generator.
-
-## Why yet another arguments parser?
-Because I haven't found such a nice one yet. :smile:
-
-Well, I liked the [Qt command line parser](http://doc.qt.io/qt-5/qcommandlineparser.html), and I wanted to create an equivalent written in NodeJS. The code looks better and more organized in this style. Also, with *clp* you can easily generate and display help content and version information.
 
 ## Installation
 
@@ -14,7 +25,8 @@ Well, I liked the [Qt command line parser](http://doc.qt.io/qt-5/qcommandlinepar
 $ npm i clp
 ```
 
-## Examples
+## Example
+
 ```js
 #!/usr/bin/env node
 
@@ -41,7 +53,6 @@ parser.addExample("name-age -a 5 # will default the name to \"Alice\"");
 
 parser.process();
 
-
 // Validate the age
 if (isNaN(parseInt(ageOption.value)) || ageOption.value < 0) {
     return console.error("Invalid age.");
@@ -54,92 +65,10 @@ if (!nameOption.value) {
 
 // Use the values
 console.log(nameOption.value + " is " + ageOption.value + " year old.");
-```
 
-The generated help content looks like this:
-
-```sh
-$ ./name-age -h
-Usage: name-age [options]
-
-Options:
-  -n, --name <name>  Someone's name
-  -a, --age <age>    Someone's age
-  -h, --help         Displays this help.
-  -v, --version      Displays version information.
-
-Examples:
-  name-age -a 10 --name Bob
-  name-age -a 5 # will default the name to "Alice"
-
-These are some final notes.
-
-Documentation can be found at https://github.com/IonicaBizau/node-clp
-```
-
-And version...
-
-```sh
-$ ./name-age -v
-name-age v1.0
-```
-
-And passing the options:
-
-```sh
-$ ./name-age -a 6
-Alice is 6 year old.
-$ ./name-age -n Bob -a 30
-Bob is 30 year old.
 ```
 
 ## Documentation
-### `CLP.Option(aliases, description, name, def)`
-Creates a new `CLPOption` instance.
-
-Usages:
-
-```js
-CLP.Option(["age", "a"], "The age value.", "age", 20);
-CLP.Option("age", "The age value.", "age", 20);
-CLP.Option({
-    aliases: ["age", "a"]
-  , description: "The age value."
-  , name: "age"
-  , def: 20
-  , handler: function (opt) {
-       // Do something with opt
-    }
-});
-```
-
-#### Params
-- **Array|Object** `aliases`: An array of strings representing the aliases  (e.g. `["name", "n"]`), a string representing a single alias (e.g. `"name"`)
- or an object containing the following fields:
-
- - `aliases` (Array): An array of strings representing the
-    aliases (e.g. `["name", "n"]`)
- - `def` (Anything): The default value.
- - `description` (String): The option description.
- - `name` (String): The option name. If provided, the parser will expect a value otherwise
-   will return or display an error.
- - `handler` (Function): The option handler which will be called when the
-   option is found in the arguments. The first parameter is the option
-   object and the scope is the `CLP` instance.
-
-- **String** `description`: The option description.
-- **String** `name`: The option name.
-- **Anything** `def`: The default value.
-
-#### Return
-- **CLPOption** An object containing the following fields:  - `aliases` (Array): An array of strings containing the computed aliases,
-    the single letter being the first ones (e.g. `["-n", "--name"]`).
- - `value` (null|String|DefaultValue): The option value which was found
-    after processing the arguments.
- - `def` (Anything): The provided default value.
- - `description` (String): The option description.
- - `name` (String): The option name.
- - `is_provided` (Boolean): A flag if the option was or not been provided.
 
 ### `CLP(args, options, clpOptions)`
 Creates a new `CLP` (command line parser) instance.
@@ -172,7 +101,6 @@ var parser = new CLP("some command", ...); // pass a command string instead of a
  - `notes` (String): Final notes placed between examples and documentation
    url in help content (default: `""`).
  - `examples` (String|Array): A string or an array of string containing examples.
-
 - **Array** `clpOptions`:
 
 #### Return
@@ -246,16 +174,16 @@ Returns the version information.
 #### Return
 - **String** The version information.
 
-
 ## How to contribute
-1. File an issue in the repository, using the bug tracker, describing the
-   contribution you'd like to make. This will help us to get you started on the
-   right foot.
-2. Fork the project in your account and create a new branch:
-   `your-great-feature`.
-3. Commit your changes in that branch.
-4. Open a pull request, and reference the initial issue in the pull request
-   message.
+Have an idea? Found a bug? See [how to contribute][contributing].
 
 ## License
-See the [LICENSE](./LICENSE) file.
+[KINDLY][license] © [Ionică Bizău][website]–The [LICENSE](/LICENSE) file contains
+a copy of the license.
+
+[license]: http://ionicabizau.github.io/kindly-license/?author=Ionic%C4%83%20Biz%C4%83u%20%3Cbizauionica@gmail.com%3E&year=2015
+[contributing]: /CONTRIBUTING.md
+[website]: http://ionicabizau.net
+[docs]: /DOCUMENTATION.md
+[paypal-donations]: https://www.paypal.com/cgi-bin/webscr?cmd=_s-xclick&hosted_button_id=MG98D7NPFZ3MG
+[donate-now]: http://i.imgur.com/jioicaN.png
